modules

+++ src/Concerns/HasService.php
@@ -9,43 +9,53 @@ trait HasService
     protected static function bootHasService()
     {
         static::created(function ($query) {
-            $service_parent = static::parentService($query->parent_id);
-            $parent_id = null;
-            if ($service_parent) $parent_id = $service_parent->getKey();
-
-            $service = $query->service()->updateOrCreate([
-                'parent_id'      => $parent_id,
-                "reference_id"   => $query->id,
-                "reference_type" => $query->getMorphClass()
-            ], [
-                'name' => $query->name
-            ]);
-
-            if (isset($query->service_code)) {
-                $service->service_code = $query->service_code;
-                // $service->price        = request()->price;
-                $service->save();
+            if ($query->isUsingService()){
+                $service_parent = static::parentService($query->parent_id);
+                $parent_id = null;
+                if ($service_parent) $parent_id = $service_parent->getKey();
+    
+                $service = $query->service()->updateOrCreate([
+                    'parent_id'      => $parent_id,
+                    "reference_id"   => $query->id,
+                    "reference_type" => $query->getMorphClass()
+                ], [
+                    'name' => $query->name
+                ]);
+    
+                if (isset($query->service_code)) {
+                    $service->service_code = $query->service_code;
+                    // $service->price        = request()->price;
+                    $service->save();
+                }
             }
         });
         static::deleting(function ($query) {
-            $query->service()->delete();
-        });
-        static::updated(function ($query) {
-            $service_parent = static::parentService($query->parent_id);
-            $parent_id = null;
-            if ($service_parent) $parent_id = $service_parent->getKey();
-            $service = $query->service()->updateOrCreate([
-                'parent_id'         => $parent_id,
-                "reference_id"      => $query->id,
-                "reference_type"    => $query->getMorphClass()
-            ], [
-                'name' => $query->name,
-            ]);
-            if ($query->service_code) {
-                $service->service_code = $query->service_code;
-                $service->save();
+            if ($query->isUsingService()){
+                $query->service()->delete();
             }
         });
+        static::updated(function ($query) {
+            if ($query->isUsingService()){
+                $service_parent = static::parentService($query->parent_id);
+                $parent_id = null;
+                if ($service_parent) $parent_id = $service_parent->getKey();
+                $service = $query->service()->updateOrCreate([
+                    'parent_id'         => $parent_id,
+                    "reference_id"      => $query->id,
+                    "reference_type"    => $query->getMorphClass()
+                ], [
+                    'name' => $query->name,
+                ]);
+                if ($query->service_code) {
+                    $service->service_code = $query->service_code;
+                    $service->save();
+                }
+            }
+        });
+    }
+
+    protected function isUsingService(): bool{
+        return true;
     }
 
     protected static function parentService($parent_id)
